refactor(core): split the checkpoint module along its lifecycle
Pure code movement; no functional change. checkpoint/mod.rs had grown to
4,500 lines covering every concern from SST segmentation to retention. It
is now a 45-line module root over nine focused submodules that follow the
manifest lifecycle: runs (LSM run model and layout policy), create
(checkpoint orchestration), build (SST writing), publish (manifest objects
and head CAS), load (read-side reconstruction), validate (descriptor and
row checks), scan (verified row scans), retention (floor advancement), and
error. Tests move to checkpoint/tests.rs unchanged.

Internal items crossing the new file boundaries become pub(super); every
path that other modules used (crate::checkpoint::*) is preserved by
re-exports from the module root, so no caller changes.
